feat-auction-schedular to inactive

+++ src/main/java/com/group1/ecocredit/dto/DisplayBidRequest.java
@@ -12,5 +12,5 @@ public class DisplayBidRequest {
 
     private Long bidId;
     @Nullable
-    private Integer newBidAmount;
+    private Double newBidAmount;
 }
